audio-rep-networks/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.2.0
ffmpeg==1.4
ffprobe==0.5
filelock==3.17.0
flake8==7.1.2
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
submitit==1.5.2
sympy==1.13.3
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.2.2
torchmetrics==1.6.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
